+++ hw4/task4.ts
@@ -1,0 +1,12 @@
+// Task 4 💻
+
+//Дана произвольная строка. Реализуйте метод, который проверяет заканчивается ли предложение точкой или нет.
+// Если оканчивается, вывести предложение - `Данное предложение закончено`, если нет, то `В конце предложения не хватает точки`
+
+const newStr4 = "This day is grate!";
+const lastSymbol = newStr4[newStr4.length-1];
+
+if(lastSymbol === ".")
+    console.log("Данное предложение закончено");
+    else
+    console.log("В конце предложения не хватает точки");
