+++ hw4/task6.ts
@@ -1,0 +1,6 @@
+// Task 6 💻
+
+//Дана строка 'я учу typescript!'. С помощью метода split запишите каждое слово этой строки в отдельный элемент массива.
+
+const str6 = "я учу typescript!"
+const arr = str6.split(' '); //Здесь получится, что последнее слово typescript! Их возможно тоже нужно разделить?
