+++ hw4/task7.ts
@@ -1,0 +1,7 @@
+// Task 7 💻
+
+//Дана строка '                я учу typescript!                '. Выведите строку, без начальных и пробелов в конце строки.
+
+const str7 = "                я учу typescript!                ";
+
+console.log(str7.trim());
