+++ hw4/task8.ts
@@ -1,0 +1,7 @@
+// Task 8 💻
+
+//Дано число 8.829734872948. Выведите число с 3 значащами цифрами после точки
+
+const number8 = 8.829734872948;
+
+console.log(number8.toFixed(3));
